requirements.txt
tensorflow==2.17.0
numpy==1.26.4
streamlit==1.39.0
Pillow==10.4.0
